% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ltns}
\alias{ltns}
\title{Chicago crime tensor dataset}
\format{
An order-3 tensor with entries representing the log counts of crimes from 24 hours, 77 community areas, and 32 crime types.
}
\source{
\url{http://frostt.io/tensors/chicago-crime/}
}
\usage{
ltns
}
\description{
Chicago crime dataset consists of crime counts reported in the city of Chicago, ranging from January 1st, 2001 to December 11th, 2017.
}
\keyword{datasets}
