% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mode_info}
\alias{mode_info}
\title{A list of mode information of the Chicago crime tensor dataset}
\format{
A list consisting of crime areas, crime hours, and crime types:
\describe{
  \item{\code{hour_map}}{24 hours of crimes}
  \item{\code{area_map}}{77 areas of crimes}
  \item{\code{crimetype_map}}{32 types of crimes}
}
}
\source{
\url{http://frostt.io/tensors/chicago-crime/}
}
\usage{
mode_info
}
\description{
A list of mode information of order-3 tensor dataset \code{ltns}.
}
\keyword{datasets}
